Add files via upload
This is the audio file for the poster (supposed to be) presented in Aix Prosody Summer School 2016. Due to the limit of space on a poster, this poster presentation doesn't include all the details that are supposed to be in the full-size presentation. Please contact the author for more information. All comments are welcome.